♻️[Refactor] editor 폼 분리 및 write 공통 로직 상위로 이동
