+++ src/components/features/record/editor/lib/mapRecordWriteToCreateBoardInput.ts
